moved visual tests to notebook
