+++ src/local_pathfinding/test/test_obstacles.py
@@ -13,8 +13,8 @@ from custom_interfaces.msg import (
 from shapely.geometry import MultiPolygon, Point, Polygon
 from shapely.strtree import STRtree
 
-from land.land_polygon_etl import load_pkl
 from local_pathfinding.coord_systems import XY, latlon_to_xy, meters_to_km
+from local_pathfinding.land_polygon_etl import load_pkl
 from local_pathfinding.obstacles import BOAT_BUFFER, Boat, Land, Obstacle
 
 SPATIAL_INDEX = load_pkl("/workspaces/sailbot_workspace/src/local_pathfinding/land/pkl/sindex.pkl")
@@ -644,129 +644,3 @@ def test_update_reference_point(
 
     # There is some error in the latlon_to_xy conversion but the results are close
     assert translation == pytest.approx(displacement, rel=0.1), "incorrect translation"
-
-
-if __name__ == "__main__":
-    """VISUAL TESTS
-
-    TODO: verify calculate_projected_distance via numerical/approximation method and show it
-    converges to the analytical solution given by calculate_projected_distance. Maybe turn into
-    an animation.
-
-    The collision zone length can be verified visually, using the plotly chart below.
-
-    The values for the measured and calculated/expected length of the collision zones
-    match, shown in the top right.
-
-    The invalid state point is within the collision zone and the valid state point is outside.
-    Validity for the same points is checked above in: test_is_valid
-
-    Increasing the cog of the ais_ship, starting from zero, corresponds to a clockwise rotation,
-    starting from true north (the y axis) as expected.
-    """
-    import plotly.graph_objects as go
-    from numpy import ndarray
-
-    # Sample AIS SHIP message
-    ais_ship = HelperAISShip(
-        id=1,
-        lat_lon=HelperLatLon(latitude=51.97917631092298, longitude=-137.1106454702385),
-        cog=HelperHeading(heading=0.0),
-        sog=HelperSpeed(speed=18.52),
-        width=HelperDimension(dimension=20.0),
-        length=HelperDimension(dimension=100.0),
-        rot=HelperROT(rot=0),
-    )
-
-    # Create a boat object
-    boat1 = Boat(
-        HelperLatLon(latitude=52.268119490007756, longitude=-136.9133983613776),
-        HelperLatLon(latitude=51.95785651405779, longitude=-136.26282894969611),
-        30.0,
-        ais_ship,
-    )
-
-    # Choose some states for visual inspection
-    valid_state = HelperLatLon(latitude=50.42973337261916, longitude=-134.12018940923838)
-    invalid_state = HelperLatLon(latitude=52.174842845359755, longitude=-137.10372451905042)
-
-    # Extract coordinates for sailbot
-    sailbot_x, sailbot_y = boat1.sailbot_position
-    sailbot = go.Scatter(x=[sailbot_x], y=[sailbot_y], mode="markers", name="Sailbot Position")
-
-    fig1 = go.Figure(sailbot)
-
-    # Extract coordinates for valid and invalid states
-    valid_state_x, valid_state_y = latlon_to_xy(boat1.reference, valid_state)
-    valid_state = go.Scatter(
-        x=[valid_state_x], y=[valid_state_y], mode="markers", name="Valid State"
-    )
-
-    fig1.add_trace(valid_state)
-
-    invalid_state_x, invalid_state_y = latlon_to_xy(boat1.reference, invalid_state)
-    invalid_state = go.Scatter(
-        x=[invalid_state_x], y=[invalid_state_y], mode="markers", name="Invalid State"
-    )
-
-    fig1.add_trace(invalid_state)
-
-    # Extract exterior coordinates for boat1's collision cone
-    if boat1.collision_zone is not None:
-        boat_x, boat_y = np.array(boat1.collision_zone.exterior.coords.xy)
-        boat_x = np.array(boat_x)
-        boat_y = np.array(boat_y)
-        boat = go.Scatter(x=boat_x, y=boat_y, fill="toself", name="Boat Collision Cone")
-        fig1.add_trace(boat)
-
-    # Manually calculate the length of the collision zone based on:
-    # - the boat's projected distance
-    # - the boat's length
-    # - the safety buffer
-    collision_zone_length = round(
-        (
-            boat1._calculate_projected_distance()
-            + 2 * BOAT_BUFFER
-            + meters_to_km(boat1.ais_ship.length.dimension)
-        ),
-        4,
-    )
-
-    fig1.add_annotation(
-        text="Calculated Length of Collision Zone : " + str(collision_zone_length) + " km",
-        align="center",
-        showarrow=False,
-        xref="paper",
-        yref="paper",
-        x=1,
-        y=1,
-        bordercolor="black",
-        borderwidth=1,
-    )
-
-    # Measure the length of the collision zone based on the points of the polygon
-    x: ndarray
-    y: ndarray
-
-    if boat1.collision_zone is not None:
-        x, y = boat1.collision_zone.exterior.coords.xy
-
-    mid_1 = Point((x[1] + x[2]) / 2, (y[1] + y[2]) / 2)
-    mid_2 = Point((x[0] + x[3]) / 2, (y[0] + y[3]) / 2)
-
-    length = round(mid_1.distance(mid_2), 4)
-
-    fig1.add_annotation(
-        text="Measured Length of Collision Zone : " + str(length) + " km",
-        align="center",
-        showarrow=False,
-        xref="paper",
-        yref="paper",
-        x=0.7,
-        y=1,
-        bordercolor="black",
-        borderwidth=1,
-    )
-
-    fig1.update_layout(yaxis_range=[-200, 200], xaxis_range=[-200, 750])
-    fig1.show()
